+++ tests/publisher/test_filter.py
@@ -1,3 +1,6 @@
+import pytest
+
+from mex.common.exceptions import MExError
 from mex.common.models import MergedConsent, MergedPerson
 from mex.extractors.publisher.filter import filter_persons_with_consent
 
@@ -6,10 +9,26 @@ def test_filter_persons_with_consent(
     merged_person_list: list[MergedPerson],
     merged_consent_list: list[MergedConsent],
 ) -> None:
-    result = filter_persons_with_consent(merged_person_list, merged_consent_list)
+    result = filter_persons_with_consent(
+        merged_person_list,
+        merged_consent_list[0:1],  # only consents referencing different persons
+    )
     assert len(result) == 1
     assert result[0].model_dump(exclude_defaults=True, mode="json") == {
         "identifier": "PersonPositiveConsent",
         "fullName": ["Consent, Positive"],
         "memberOf": ["SomeUnitIdentifier"],
     }
+
+
+def test_filter_persons_with_consent__more_than_one_consent(
+    merged_person_list: list[MergedPerson],
+    merged_consent_list: list[MergedConsent],
+) -> None:
+    with pytest.raises(
+        MExError, match=r"Merged Person PersonTwoConsents has more than 1 consent."
+    ):
+        filter_persons_with_consent(
+            merged_person_list,
+            merged_consent_list,  # all consents incl. those referencing the same person
+        )